ref: Refactor project to hexagonal layout
