+++ internal/adapter/outbound/repository/local/pathvalidator/validator.go
