+++ internal/adapter/outbound/repository/local/storage.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/andreychano/compressor-golang/internal/adapter/repository/local/pathvalidator"
-	"github.com/andreychano/compressor-golang/pkg/core/domain"
+	"github.com/andreychano/compressor-golang/internal/adapter/outbound/repository/local/pathvalidator"
+	"github.com/andreychano/compressor-golang/internal/pkg/core/domain"
 )
 
 type LocalFileStorage struct {
