+++ internal/pkg/core/domain/file.go
