+++ internal/pkg/core/port/repository.go
@@ -3,7 +3,7 @@ package port
 import (
 	"context"
 
-	"github.com/andreychano/compressor-golang/pkg/core/domain"
+	"github.com/andreychano/compressor-golang/internal/pkg/core/domain"
 )
 
 // File repisitory определяет контракт для хранения файлов
